refactored policies and added deq and deqmpc policies
