+++ deqmpc/train_simple.py
@@ -8,163 +8,9 @@ import qpth.qp_wrapper as mpc
 import ipdb
 from envs import PendulumEnv, PendulumDynamics
 from datagen import get_gt_data, merge_gt_data, sample_trajectory
+from policies import NNMPCPolicy, DEQPolicy, DEQMPCPolicy
 
 ## example task : hard pendulum with weird coordinates to make sure direct target tracking is difficult
-
-
-class FFDNetwork(torch.nn.Module):
-    """
-    Feedforward network to generate reference trajectories of horizon T
-    """
-
-    def __init__(self, args, env):
-        super().__init__()
-        self.args = args
-        self.nu = env.nu
-        self.nx = env.nx
-        self.np = args.np
-        self.T = args.T
-
-        ## define the network layers :
-        self.fc1 = torch.nn.Linear(self.nx, 256)
-        self.ln1 = torch.nn.LayerNorm(256)
-        self.relu1 = torch.nn.ReLU()
-        self.fc2 = torch.nn.Linear(256, 256)
-        self.ln2 = torch.nn.LayerNorm(256)
-        self.relu2 = torch.nn.ReLU()
-        self.fc3 = torch.nn.Linear(256, self.np * self.T)
-        self.net = torch.nn.Sequential(
-            self.fc1, self.ln1, self.relu1, self.fc2, self.ln2, self.relu2, self.fc3
-        )
-
-    def forward(self, x):
-        """
-        compute the policy output for the given state x
-        """
-        dx_ref = self.net(x)
-        dx_ref = dx_ref.view(-1, self.T, self.np)
-        x_ref = dx_ref + x[:, None, : self.np]
-        return x_ref
-
-
-class Tracking_MPC(torch.nn.Module):
-    def __init__(self, args, env):
-        super().__init__()
-        self.args = args
-        self.nu = env.nu
-        self.nx = env.nx
-        self.dt = env.dt
-        self.T = args.T
-
-        # May comment out input constraints for now
-        self.u_upper = None  # torch.tensor(env.action_space.high).to(args.device)
-        self.u_lower = None  # torch.tensor(env.action_space.low).to(args.device)
-        self.qp_iter = args.qp_iter
-        self.eps = args.eps
-        self.warm_start = args.warm_start
-        self.bsz = args.bsz
-        self.device = args.device
-
-        self.Q = args.Q
-        self.R = args.R
-        # self.Qf = args.Qf
-        if args.Q is None:
-            self.Q = torch.ones(self.nx, dtype=torch.float32, device=args.device)
-            # self.Qf = torch.ones(self.nx, dtype=torch.float32, device=args.device)
-            self.R = torch.ones(self.nu, dtype=torch.float32, device=args.device)
-        self.Q = torch.cat([self.Q, self.R], dim=0)
-        self.Q = torch.diag(self.Q).repeat(self.T, self.bsz, 1, 1)
-
-        self.u_init = torch.zeros(
-            self.T, self.bsz, self.nu, dtype=torch.float32, device=self.device
-        )
-
-        self.ctrl = mpc.MPC(
-            self.nx,
-            self.nu,
-            self.T,
-            u_lower=self.u_lower,
-            u_upper=self.u_upper,
-            qp_iter=self.qp_iter,
-            exit_unconverged=False,
-            eps=1e-2,
-            n_batch=self.bsz,
-            backprop=False,
-            verbose=0,
-            u_init=self.u_init,
-            grad_method=mpc.GradMethods.AUTO_DIFF,
-            solver_type="dense",
-        )
-
-    def forward(self, x_init, x_ref):
-        """
-        compute the mpc output for the given state x and reference x_ref
-        """
-
-        self.compute_p(x_ref)
-        cost = mpc.QuadCost(self.Q, self.p)
-        self.ctrl.u_init = self.u_init
-        state = x_init  # .unsqueeze(0).repeat(self.bsz, 1)
-        nominal_states, nominal_actions = self.ctrl(state, cost, PendulumDynamics())
-        return nominal_states, nominal_actions
-
-    def compute_p(self, x_ref):
-        """
-        compute the p for the quadratic objective using self.Q as the diagonal matrix and the reference x_ref at each time without a for loop
-        """
-        self.p = torch.zeros(
-            self.T, self.bsz, self.nx + self.nu, dtype=torch.float32, device=self.device
-        )
-        self.p[:, :, : self.nx] = -(
-            self.Q[:, :, : self.nx, : self.nx] * x_ref.unsqueeze(-2)
-        ).sum(dim=-1)
-        return self.p
-
-
-class NNMPCPolicy(torch.nn.Module):
-    def __init__(self, args, env):
-        super().__init__()
-        self.args = args
-        self.nu = env.nu
-        self.nx = env.nx
-        self.np = args.np
-        self.T = args.T
-        self.dt = env.dt
-        self.device = args.device
-        self.model = FFDNetwork(args, env)
-        self.model.to(self.device)
-        self.tracking_mpc = Tracking_MPC(args, env)
-
-    def forward(self, x):
-        """
-        compute the policy output for the given state x
-        """
-        x_ref = self.model(x)
-        # x_ref = x_ref.view(-1, self.np)
-        x_ref = torch.cat(
-            [
-                x_ref,
-                torch.zeros(
-                    list(x_ref.shape[:-1])
-                    + [
-                        self.np,
-                    ]
-                ).to(self.args.device),
-            ],
-            dim=-1,
-        ).transpose(0, 1)
-        nominal_states, nominal_actions = self.tracking_mpc(x, x_ref)
-        return nominal_states, nominal_actions
-
-
-# class DEQPolicy:
-
-# class DEQMPCPolicy:
-
-# class NNPolicy:
-
-# class NNMPCPolicy:
-
 
 def main():
     import argparse
